Make sure auth provider listener is setup before extensions and move register warning into ext host code (#233373)
This just moves code out of a contribution into the service so that the ExtPoint handler gets set up before extensions start coming in.

Also, we'd like to throw when an extension doesn't include their auth provider in the package.json so move the warning into a place that makes sense and set up a telemetry event so we can understand usage of this behavior.

Fixes https://github.com/microsoft/vscode/issues/233324
